separate out appStep api
- the notion of an appStep is now a logStep
- appSteps call log step
- separate consumable appSteps into api/ from things appSteps use until
  util/
- prefer split files
- prefer export default where possible
- create an index.js that supplies the most common steps; the rest of
  the API can still be imported as needed
- remove various holdover steps that are redundant or not used
- update tests

The two-hierarchy approach as seen in Zealot isn't implemented here. I
figure createTestBrim can have that, if we ever want it.
