+++ itest/lib/appStep/api/index.js
@@ -1,0 +1,22 @@
+/* @flow */
+
+import {click, rightClick} from "./click"
+import ingestFile from "./ingestFile"
+import reload from "./reload"
+import savePcap from "./savePcap"
+import search from "./search"
+import setSpan from "./setSpan"
+import startApp from "./startApp"
+import takeScreenshot from "./takeScreenshot"
+
+export default {
+  click: click,
+  rightClick: rightClick,
+  ingestFile: ingestFile,
+  reload: reload,
+  savePcap: savePcap,
+  search: search,
+  setSpan: setSpan,
+  startApp: startApp,
+  takeScreenshot: takeScreenshot
+}
